requirements-jax-cuda.txt
# Tensorflow cpu-only version (needed for testing).
tensorflow-cpu~=2.16.2  # Pin to TF 2.16

# Torch cpu-only version (needed for testing).
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.1.0
torchvision>=0.16.0

# Jax with cuda support.
# TODO: 0.4.24 has an updated Cuda version breaks Jax CI.
--find-links https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
jax[cuda12-pip]==0.4.30
flax

-r requirements-common.txt